Change MM communicate RPC protocol of call endpoint
Replace buffer address and size parameter by offset in buffer parameter
and move to 64 bit FF-A direct message call. Deny all 32 bit direct
messages in SMM gateway.

+++ components/rpc/mm_communicate/endpoint/sp/test/mock_mm_service.h
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: BSD-3-Clause */
 /*
- * Copyright (c) 2021, Arm Limited and Contributors. All rights reserved.
+ * Copyright (c) 2021-2022, Arm Limited and Contributors. All rights reserved.
  */
 
 #ifndef MOCK_MM_SERVICE_H_
@@ -16,7 +16,7 @@ void mock_mm_service_init(void);
 
 void expect_mock_mm_service_receive(struct mm_service_interface *iface,
 				    const struct mm_service_call_req *req,
-				    int32_t result);
+				    int64_t result);
 
 int32_t mock_mm_service_receive(struct mm_service_interface *iface,
 				struct mm_service_call_req *req);
